finance (StockSim)/requirements.txt
cs50
Flask
Flask-Session
pytz
requests
werkzeug.security